requirements.txt
# ConverseAgent Dependencies

# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# LLM
anthropic>=0.40.0
openai>=1.12.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aiomysql>=0.2.0
aiosqlite>=0.20.0
alembic>=1.13.0
redis>=5.0.0

# HTTP Client
httpx>=0.26.0

# Email
aiosmtplib>=3.0.0
jinja2>=3.1.0

# Voice / Pipecat
pipecat-ai>=0.0.30
websockets>=12.0

# Logging
structlog>=24.1.0

# Retry
tenacity>=8.2.0

# Config
pyyaml>=6.0.1

# Dev / Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
